Create new ab.js
