+++ static/src/javascripts/projects/common/modules/experiments/ab-old.spec.js
@@ -1,5 +1,5 @@
 // @flow
-import { segment, forceSegment, run } from 'common/modules/experiments/ab';
+import { segment, forceSegment, run } from 'common/modules/experiments/ab-old';
 import {
     getParticipations,
     isParticipating,
